feat: lagt til postgres
