+++ cmd/postgresCmd/users.go
@@ -1,0 +1,15 @@
+package postgresCmd
+
+import "github.com/urfave/cli/v2"
+
+func usersCommand() *cli.Command {
+	return &cli.Command{
+		Name:    "users",
+		Aliases: []string{"u"},
+		Usage:   "Manage users in your Postgres instance",
+		Subcommands: []*cli.Command{
+			usersAddCommand(),
+			usersListCommand(),
+		},
+	}
+}
