--- cmd/root/postgres/grant.go
@@ -1,86 +1,0 @@
-package postgres
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"io"
-	"os"
-	"os/exec"
-	"strings"
-	"time"
-
-	"github.com/nais/cli/cmd"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-)
-
-var grantCmd = &cobra.Command{
-	Use:   "grant [app-name] [flags]",
-	Short: "Grant yourself access to a Postgres database.",
-	Long: `Grant yourself access to a Postgres database.
-
-	This is done by temporarily adding your user to the list of users that can administrate Cloud SQL instances and creating a user with your email.`,
-	Args: cobra.ExactArgs(1),
-	RunE: func(command *cobra.Command, args []string) error {
-		appName := args[0]
-		namespace := viper.GetString(cmd.NamespaceFlag)
-		context := viper.GetString(cmd.ContextFlag)
-		databaseName := viper.GetString(cmd.DatabaseFlag)
-		ctx := command.Context()
-
-		dbInfo, err := NewDBInfo(appName, namespace, context, databaseName)
-		if err != nil {
-			return err
-		}
-
-		projectID, err := dbInfo.ProjectID(ctx)
-		if err != nil {
-			return err
-		}
-
-		connectionName, err := dbInfo.ConnectionName(ctx)
-		if err != nil {
-			return err
-		}
-
-		fmt.Println("Grant user access")
-		if err := grantUserAccess(ctx, projectID, "roles/cloudsql.admin", 5*time.Minute); err != nil {
-			return err
-		}
-		fmt.Println("Create sql user")
-		if err := createSQLUser(ctx, projectID, connectionName); err != nil {
-			fmt.Fprintln(os.Stderr, "Error creating SQL user. One might already exist.")
-			return err
-		}
-
-		return nil
-	},
-}
-
-func createSQLUser(ctx context.Context, projectID, instance string) error {
-	email, err := currentEmail(ctx)
-	if err != nil {
-		return err
-	}
-
-	args := []string{
-		"sql",
-		"users",
-		"create",
-		email,
-		"--instance", strings.Split(instance, ":")[2],
-		"--type", "cloud_iam_user",
-		"--project", projectID,
-	}
-
-	buf := &bytes.Buffer{}
-	cmd := exec.CommandContext(ctx, "gcloud", args...)
-	cmd.Stdout = buf
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		io.Copy(os.Stdout, buf)
-		return fmt.Errorf("error running gcloud command: %w", err)
-	}
-	return nil
-}
